refactor: move buildStarFilter to core/daily-logic for clean layering (#1208 M7)
\`dashboard-data.ts\` was importing \`buildStarFilter\` from a sibling command
module (\`daily.ts\`), creating a \`commands → commands\` dependency that the
audit (#1208) flagged as a layering violation.

\`buildStarFilter\` is pure logic over \`Readonly<AgentState>\` — no IO, no
state mutation. It belongs one layer down with the other pure helpers in
\`core/daily-logic.ts\` alongside \`toShelvedPRRef\`, \`computeRepoSignals\`,
\`assessCapacity\`, etc.

Changes:
- Moved \`buildStarFilter\` from \`commands/daily.ts\` → \`core/daily-logic.ts\`
- Re-exported from \`core/index.js\` barrel alongside \`toShelvedPRRef\`
- \`commands/daily.ts\` now imports + re-exports for backward compatibility
  with anyone using the old import path
- \`commands/dashboard-data.ts\` updated to import from \`../core/index.js\`
  (the canonical layer) instead of \`./daily.js\`
- \`dashboard-data.test.ts\` mock updated to override \`buildStarFilter\` and
  \`toShelvedPRRef\` on the new import path

No runtime behavior change. Lint, typecheck, 2,725 tests pass.
