+++ packages/core/src/core/index.ts
@@ -74,6 +74,7 @@ export {
   collectActionableIssues,
   computeActionMenu,
   toShelvedPRRef,
+  buildStarFilter,
   formatActionHint,
   formatBriefSummary,
   formatSummary,
